+++ src/Git/Plantation/Env.hs
@@ -11,6 +11,7 @@ import           Data.Extensible
 import qualified Drone.Client          as Drone
 import           Git.Plantation.Config
 import qualified GitHub.Auth           as GitHub
+import           Shelly                hiding (FilePath)
 
 type Plant = RIO Env
 
@@ -24,3 +25,15 @@ type Env = Record
 
 instance HasLogFunc Env where
   logFuncL = lens (view #logger) (\x y -> x & #logger `set` y)
+
+maybeWithLogError :: Maybe a -> Text -> Plant a
+maybeWithLogError (Just x) _ = pure x
+maybeWithLogError Nothing e  = logError (display e) >> fail (show e)
+
+shelly' :: Sh a -> Plant a
+shelly' sh = do
+  env <- ask
+  shelly
+    $ (log_stdout_with (runRIO env . logDebug . display))
+    $ (log_stderr_with (runRIO env . logDebug . display))
+    $ sh