+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog for git-plantation
 
 ## Unreleased changes
+
+* org アカウント以外で config 設定 (#10)
+* dotenv ファイルが使えるようになる (#10)
+* `/score` エンドポイントで drone から取得できなくても空リストを返す (#10)